chatbot_client/requirements.txt
streamlit
openai
python-dotenv
mcp
nest_asyncio
uv